Fixed weapon variation.  Sniper rifle now spawns.  Finalized stats for 'Standard' weapon.
The concept of 'weapon progress' was not implemented and has been removed, so all weapon stats rolled as 0 - 100 before converted to the min / max ranges in the Info script for each weapon type.  Receiver type was beiing rolled from 0 to 3.0 (3.0 = sniper) so the odds of getting a sniper rifle were very, very small.  Now all weapons have equal probability.  Finally, Standard has been adjusted / play tested by me.  Feel free to play test more and adjust the damage amount in the prefab, or the other stat values in 'Projectile Standard'.  If editing that file, the left value is the 'worst' and the right value is the 'best'. For values like recoil, less is better!